requirements.txt
boto3>=1.34.0
botocore>=1.34.0
pyyaml>=6.0.1
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
